python/openai-agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anima-toolkit-openai"
version = "0.1.0"
description = "Anima tools for the OpenAI Agents SDK"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "anima-labs>=0.1.0",
    "openai-agents>=0.1.0",
]

[project.urls]
Homepage = "https://docs.anima.email"
Repository = "https://github.com/anima-labs-ai/toolkit"

[tool.hatch.build.targets.wheel]
packages = ["src/anima_toolkit_openai"]
